Move ImportSummary to db-backed ImportSession.
The rationale was that it would probably be good to keep data about the
import session around, such was the original file imported and a record
of each line imported.

I've also added a ImportedPurchase model to map ImportSessions to
Purchases and keep the original line numbers around, should an UI item
be added to allow comparisons between what is in the DB and what was in
the file.

An incipient Controller for ImportSession was added, but is yet to be
hooked up.